turn annotation settings into a popover
